Support SNS auto-configuration in non-web environments resolves #73
Prior to this commit the SNS support has been only enabled if Spring Web MVC classes
where available in the class path. With this commit, the SNS client is created if SNS classes
are available in the classpath. If there are Spring Web MVC classes available as well, the
argument resolvers will be configured.