Fix duplicate error logging
The AbstractMethodMessageHandler always logs an exception when there is
no handler method for a particular exception and the
SimpleMessageListenerContainer logs the same exception again when the
message deletion policy is ON_SUCCESS. This results in duplicated error
logging.

This commit fixes this issue by delegating exception processing to
AbstractMethodMessageHandler only when there is a configured handler.

Fixes: gh-394