Disable ContextInstanceDataAutoConfiguration by default.
ContextInstanceDataAutoConfiguration has a condition to run only in AWS environment. This condition is expensive to run outside of AWS environment and it slows down application startup several seconds.

This change makes `ContextInstanceDataAutoConfiguration` an opt-in feature - to use it users have to explicitly set the `cloud.aws.instance.data.enabled` to `true`.